Organize different classes into multiple packages
