+++ src/main/java/neochat/task/Event.java
@@ -1,8 +1,9 @@
-package neochat;
+package neochat.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
+import neochat.task.taskexception.EmptyTaskDescriptionException;
 
 public class Event extends Task {
     private LocalDateTime from;
