+++ src/main/java/neochat/task/taskexception/EmptyTaskDescriptionException.java
@@ -1,4 +1,4 @@
-package neochat;
+package neochat.task.taskexception;
 
 public class EmptyTaskDescriptionException extends Exception {
     public EmptyTaskDescriptionException() {
